+++ src/utils/redownloadIndex.h
@@ -1,4 +1,0 @@
-#pragma once
-#include <Geode/Geode.hpp>
-
-void redownloadIndex();